Made package containing shiny app
The only problem is that the path that is created and called to access the shiny app when the function run_shiny() is called is faulty.  system.file() is a function that can generate the path to the shiny application (inst/Shiny/ShinyApp.R) however on a map there is a difference between / and \ and this function doesn't find the file.  If anyone wants to have a go fixing this go for it!